refactor: #243 userService use properties & separate UserWithdrawService
