+++ src/main/java/com/umc/linkyou/service/users/UserService.java
@@ -38,9 +38,6 @@ public interface UserService {
     // 임시 비밀번호 전송
     void sendTempPassword(String email);
 
-    Users withdrawUser(Long userId, UserRequestDTO.DeleteReasonDTO deleteReasonDTO);
-
     UserResponseDTO.TokenPair reissueRefreshToken(String refreshToken);
 
-    void testImmediateDelete(Long userId);
 }
